Move contents of LH5Source.h to the source
